KUBIG CONTEST/CV_Soccer/requirements.txt
# requirements
ffmpeg-python
numpy
tensorflow
keras
scipy
resampy
torch
torchaudio
torchvision
torchvggish
tqdm
scikit-learn
opencv-python
matplotlib